Add per-provider background_analysis toggle via a new FileSystemProvider base class
Introduces a background_analysis config option so proactive nightly analysis
scanning can be disabled per filesystem provider instance. It defaults False
on the shared base and True only on LocalFileSystemProvider (inherited by
SMB/NFS, which mount to a real local path and reuse its I/O), while
webdav/cloud-storage providers keep the safe False default since analysis
would stream the full track over the network.

Splits the shared sync/parsing logic out of LocalFileSystemProvider into a
new FileSystemProvider base class in filesystem_local/base.py, since that
class was previously doing double duty as both the base every filesystem
provider subclassed and the concrete "plain local disk" implementation -
LocalFileSystemProvider is now a thin subclass of it.

Also fixes the audio-analysis background-scan gate itself: it checked
isinstance(provider, LocalFileSystemProvider) but never actually read the
new config value, and filtered candidate tracks by provider_domain rather
than provider_instance, which would incorrectly mix settings across multiple
instances of the same provider domain.
