+++ music_assistant/providers/filesystem_smb/__init__.py
@@ -21,16 +21,6 @@ from music_assistant.providers.filesystem_local import (
     ismount,
     makedirs,
 )
-from music_assistant.providers.filesystem_local.constants import (
-    CONF_CONTENT_TYPE,
-    CONF_ENTRY_IGNORE_ALBUM_PLAYLISTS,
-    CONF_ENTRY_LIBRARY_SYNC_AUDIOBOOKS,
-    CONF_ENTRY_LIBRARY_SYNC_PLAYLISTS,
-    CONF_ENTRY_LIBRARY_SYNC_PODCASTS,
-    CONF_ENTRY_LIBRARY_SYNC_TRACKS,
-    CONF_ENTRY_MISSING_ALBUM_ARTIST,
-    CONF_ENTRY_PROPAGATE_GENRES,
-)
 
 if TYPE_CHECKING:
     from music_assistant_models.config_entries import ProviderConfig
@@ -91,11 +81,9 @@ class SMBFileSystemProvider(LocalFileSystemProvider):
 
     async def get_config_entries(self) -> tuple[ConfigEntry, ...]:
         """Return Config entries to configure this provider."""
-        # connection details and content type are collected by the setup flow; surface the
-        # (immutable) content type read-only so the sync options' depends_on chains resolve
-        content_type = str(self.get_setup_value(CONF_CONTENT_TYPE, "music"))
+        entries = await super().get_config_entries()
         return (
-            ConfigEntry(key=CONF_CONTENT_TYPE, type=ConfigEntryType.LABEL, value=content_type),
+            *entries,
             ConfigEntry(
                 key=CONF_CACHE_MODE,
                 type=ConfigEntryType.STRING,
@@ -108,13 +96,6 @@ class SMBFileSystemProvider(LocalFileSystemProvider):
                     ConfigValueOption("none"),
                 ],
             ),
-            CONF_ENTRY_MISSING_ALBUM_ARTIST,
-            CONF_ENTRY_IGNORE_ALBUM_PLAYLISTS,
-            CONF_ENTRY_LIBRARY_SYNC_TRACKS,
-            CONF_ENTRY_LIBRARY_SYNC_PLAYLISTS,
-            CONF_ENTRY_LIBRARY_SYNC_PODCASTS,
-            CONF_ENTRY_LIBRARY_SYNC_AUDIOBOOKS,
-            CONF_ENTRY_PROPAGATE_GENRES,
         )
 
     async def handle_async_init(self) -> None:
