+++ tests/providers/filesystem_cloud/test_base.py
@@ -465,7 +465,7 @@ def _make_playlist_provider(
     provider = _make_provider(tree=PLAYLIST_TREE)
     # the inherited implementation lives in (and imports from) filesystem_local
     monkeypatch.setattr(
-        "music_assistant.providers.filesystem_local.async_parse_tags",
+        "music_assistant.providers.filesystem_local.base.async_parse_tags",
         AsyncMock(return_value=MagicMock()),
     )
     track = MagicMock(name="track")
